feat: basic User CRUD

+++ src/main/java/com/lsoria/qrsec/controller/rest/UserController.java
@@ -1,11 +1,13 @@
 package com.lsoria.qrsec.controller.rest;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import com.lsoria.qrsec.domain.dto.UserDTO;
 import com.lsoria.qrsec.domain.dto.mapper.UserMapper;
+import com.lsoria.qrsec.domain.model.Role;
 import com.lsoria.qrsec.domain.model.User;
 import com.lsoria.qrsec.service.UserService;
 import com.lsoria.qrsec.service.exception.ConflictException;
@@ -32,6 +34,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -48,7 +51,17 @@ public class UserController {
     UserMapper userMapper;
 
     @Operation(summary = "Get all Users (privileged)", description = "Get all Users from the neighbourhood")
-    @GetMapping("${api.path.users}")
+    @GetMapping("${api.path.admin.users}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the Admin that wants to see the Users",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
@@ -66,22 +79,71 @@ public class UserController {
                     content = @Content()
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized (not a privileged User)",
+                    content = @Content()
+            ),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "User not found on the database",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "500",
                     description = "Some error prevented the Users from being retrieved",
                     content = @Content()
             )
     })
-    // TODO: Add @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<List<UserDTO>> getUsers() {
+    public ResponseEntity<List<UserDTO>> getUsers(
+            @RequestHeader(value = "X-Email") @NotNull String email
+    ) {
 
-        List<User> users = userService.findAll();
+        try {
 
-        return ResponseEntity.ok(users.stream().map(userMapper::userToUserDTO).collect(Collectors.toList()));
+            // TODO: Add @PreAuthorize("hasAuthority('ADMIN')")
+            if (!userService.userIsAuthorized(email, new Role(Role.ADMIN))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
+
+            List<User> users = userService.findAll();
+            if (users.isEmpty()) {
+
+                return ResponseEntity.noContent().build();
+
+            }
+
+            return ResponseEntity.ok(users.stream().map(userMapper::userToUserDTO).collect(Collectors.toList()));
+
+        } catch (NotFoundException exception) {
+
+            log.error("Message: {}.", exception.getMessage());
+
+            return ResponseEntity.notFound().build();
+
+        } catch (Exception exception) {
+
+            log.error("Couldn't find all the Users.\nMessage: {}.\nStackTrace:\n{}", exception.getMessage(), exception.getStackTrace());
+
+        }
+
+        return ResponseEntity.internalServerError().build();
 
     }
 
-    @Operation(summary = "Get a User (privileged)", description = "Get an specific User")
+    @Operation(summary = "Get a User (privileged, guard or self)", description = "Get an specific User")
     @GetMapping("${api.path.users}/{id}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the User that wants to see a User",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @Parameter(
             name = "id",
             description = "User uuid",
@@ -103,6 +165,11 @@ public class UserController {
                     )
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "404",
                     description = "User not found",
                     content = @Content()
@@ -113,20 +180,49 @@ public class UserController {
                     content = @Content()
             )
     })
-    // TODO: Add @PreAuthorize("hasAuthority('ADMIN')")
     public ResponseEntity<UserDTO> getUser(
+            @RequestHeader(value = "X-Email") @NotNull String email,
             @PathVariable @NotNull String id
     ) {
 
-        Optional<User> user = userService.findOne(id);
+        try {
 
-        if (user.isEmpty()) {
-            return ResponseEntity.notFound().build();
+            // TODO: Replace with @PreAuthorize("hasAuthority('OWNER') or hasAuthority('ADMIN') or hasAuthority('GUARD')")
+            Optional<User> currentUser = userService.findByUsername(email);
+            if (currentUser.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+            Optional<User> user = userService.findOne(id);
+            if (user.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+            if ((currentUser.get().getAuthorities() == null || currentUser.get().getAuthorities().isEmpty()) || (currentUser.get().getAuthorities().contains(new Role(Role.OWNER)) && !Objects.equals(currentUser.get().getId(), user.get().getId()))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
+
+            if (!userService.userIsAuthorized(email, new Role(Role.ADMIN)) && !user.get().isEnabled()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+
+            UserDTO userDTO = userMapper.userToUserDTO(user.get());
+
+            return ResponseEntity.ok(userDTO);
+
+        } catch (Exception exception) {
+
+            log.error("Couldn't find the Guest {}.\nMessage: {}.\nStackTrace:\n{}", id, exception.getMessage(), exception.getStackTrace());
+
         }
 
-        UserDTO userDTO = userMapper.userToUserDTO(user.get());
-
-        return ResponseEntity.ok(userDTO);
+        return ResponseEntity.internalServerError().build();
 
     }
 
@@ -164,11 +260,11 @@ public class UserController {
             @RequestBody @NotNull UserDTO userDTO
     ) {
 
-        User userToCreate = userMapper.userDTOToUser(userDTO);
-        userToCreate.setId(null);
-        userToCreate.setEnabled(false);
-
         try {
+
+            User userToCreate = userMapper.userDTOToUser(userDTO);
+            userToCreate.setId(null);
+            userToCreate.setEnabled(false);
 
             User createdUser = userService.save(userToCreate);
 
@@ -176,13 +272,13 @@ public class UserController {
 
         } catch (ConflictException conflictException) {
 
-            log.error("Couldn't create USER.\nMessage: {}.", conflictException.getMessage());
+            log.error("Message: {}.", conflictException.getMessage());
 
             return ResponseEntity.status(HttpStatus.CONFLICT).build();
 
         } catch (Exception exception) {
 
-            log.error("Couldn't create USER with values:\n{}\nMessage: {}.\nStackTrace:\n{}", userDTO, exception.getMessage(), exception.getStackTrace());
+            log.error("Couldn't create User with values:\n{}\nMessage: {}.\nStackTrace:\n{}", userDTO, exception.getMessage(), exception.getStackTrace());
 
         }
 
@@ -190,8 +286,18 @@ public class UserController {
 
     }
 
-    @Operation(summary = "Update a User", description = "Update user's information")
+    @Operation(summary = "Update a User (privileged or self)", description = "Update user's information")
     @PutMapping("${api.path.users}/{id}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the User that wants to update the User",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @Parameter(
             name = "id",
             description = "User uuid",
@@ -221,6 +327,11 @@ public class UserController {
                     )
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "404",
                     description = "User not found",
                     content = @Content()
@@ -232,16 +343,29 @@ public class UserController {
             )
     })
     public ResponseEntity<UserDTO> updateUser(
+            @RequestHeader(value = "X-Email") @NotNull String email,
             @PathVariable @NotNull String id,
             @RequestBody @NotNull UserDTO userDTO
     ) {
 
         try {
 
+            // TODO: Replace with @PreAuthorize("hasAuthority('OWNER') or hasAuthority('ADMIN')")
+            Optional<User> currentUser = userService.findByUsername(email);
+            if (currentUser.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
             Optional<User> foundUser = userService.findOne(id);
             if (foundUser.isEmpty()) {
 
                 return ResponseEntity.notFound().build();
+
+            }
+            if (currentUser.get().getAuthorities().contains(new Role(Role.ADMIN)) || (currentUser.get().getAuthorities().contains(new Role(Role.OWNER)) && !Objects.equals(currentUser.get().getId(), foundUser.get().getId()))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
             }
 
@@ -254,7 +378,7 @@ public class UserController {
 
         } catch (Exception exception) {
 
-            log.error("Couldn't update USER with values:\n{}\nMessage: {}.\nStackTrace:\n{}", userDTO, exception.getMessage(), exception.getStackTrace());
+            log.error("Couldn't update User with values:\n{}\nMessage: {}.\nStackTrace:\n{}", userDTO, exception.getMessage(), exception.getStackTrace());
 
         }
 
@@ -262,8 +386,18 @@ public class UserController {
 
     }
 
-    @Operation(summary = "Delete a User", description = "The user will no longer be available in the database")
+    @Operation(summary = "Delete a User (privileged or self)", description = "The user will be disabled")
     @DeleteMapping("${api.path.users}/{id}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the User that wants to delete it's account",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @Parameter(
             name = "id",
             description = "User uuid",
@@ -282,21 +416,40 @@ public class UserController {
                     content = @Content()
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "404",
                     description = "User not found",
                     content = @Content()
             ),
             @ApiResponse(
                     responseCode = "500",
-                    description = "Some error prevented the User from being deleted or the Owner from being removed", // TODO: Remove all references of the user from Guests
+                    description = "Some error prevented the User from being deleted or the Owner from being removed",
                     content = @Content()
             )
     })
     public ResponseEntity<UserDTO> deleteUser(
+            @RequestHeader(value = "X-Email") @NotNull String email,
             @PathVariable @NotNull String id
     ) {
 
         try {
+
+            Optional<User> currentUser = userService.findByUsername(email);
+            if (currentUser.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+            // TODO: Replace with @PreAuthorize("hasAuthority('OWNER') or hasAuthority('ADMIN')")
+            if ((!Objects.equals(currentUser.get().getId(), id) && userService.userIsAuthorized(email, new Role(Role.OWNER))) || userService.userIsAuthorized(email, new Role(Role.GUARD))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
 
             userService.delete(id);
 
@@ -304,11 +457,13 @@ public class UserController {
 
         } catch (NotFoundException exception) {
 
+            log.error("Message: {}.", exception.getMessage());
+
             return ResponseEntity.notFound().build();
 
         } catch (Exception exception) {
 
-            log.error("Couldn't delete USER {}.\nMessage: {}.\nStackTrace:\n{}", id, exception.getMessage(), exception.getStackTrace());
+            log.error("Couldn't delete User {}.\nMessage: {}.\nStackTrace:\n{}", id, exception.getMessage(), exception.getStackTrace());
 
         }
